+++ www/src/Controller/Integration.php
@@ -36,7 +36,7 @@ class Integration extends ControllerAbstract
         $data = [
             'title' => 'Integration | TypeSchema',
             'method' => explode('::', __METHOD__),
-            'types' => array_chunk($types, ceil(count($types) / 2), true),
+            'types' => array_chunk($types, (int) ceil(count($types) / 2), true),
         ];
 
         $templateFile = __DIR__ . '/../../resources/template/integration.php';
